Extracting code utils (for easier testing)
